--- a/PlaceRouteHierFlow/router/A_star.cpp
+++ b/PlaceRouteHierFlow/router/A_star.cpp
@@ -514,7 +514,7 @@ bool A_star::CheckExendable_With_Certain_Length_Tail_Extend(int first_node_same_
 };
 
 int A_star::Calculate_Interval_number(Grid &grid, int node) {
-  auto logger = spdlog::default_logger()->clone("router.A_star.Calculate_Interval_number");
+  //auto logger = spdlog::default_logger()->clone("router.A_star.Calculate_Interval_number");
 
   int interval_number = 1;
   int metal = grid.vertices_total[node].metal;
@@ -526,17 +526,10 @@ int A_star::Calculate_Interval_number(Grid &grid, int node) {
     via_space_length = drc_info.Via_info[metal].width + drc_info.Via_info[metal].dist_ss;
     pitches = drc_info.Metal_info[metal].grid_unit_x;
     interval_number = ceil((double)via_space_length / pitches);
-    // logger->debug("metal {0} via_space_length {1} pitches {2}", metal, via_space_length, pitches);
-    // logger->debug("interval_number 1 {0}", interval_number);
-    // assert(0);
-
   } else {
     via_space_length = drc_info.Via_info[metal].width_y + drc_info.Via_info[metal].dist_ss_y;
     pitches = drc_info.Metal_info[metal].grid_unit_y;
     interval_number = ceil((double)via_space_length / pitches);
-    // logger->debug("metal {0} via_space_length {1} pitches {2}", metal, via_space_length, pitches);
-    // logger->debug("interval_number 2 {0}", interval_number);
-    // assert(0);
   }
 
   return interval_number;
@@ -697,12 +690,6 @@ bool A_star::find_succsive_parallel_node(Grid &grid, int current_node, int left,
     vector<int> temp_nodes;
     int exist = 0;
     if (mode == 0) {
-      /*
-            if(hide_mode){
-              exist = find_nodes_west(grid, current_node, left, temp_nodes);
-              exist = Check_Src_Dest(temp_nodes, src_index);
-            }
-      */
       if (!exist) {
         temp_nodes.clear();
         exist = find_nodes_south(grid, current_node, left, temp_nodes);
@@ -724,12 +711,6 @@ bool A_star::find_succsive_parallel_node(Grid &grid, int current_node, int left,
     vector<int> temp_nodes;
     int exist = 0;
     if (mode == 0) {
-      /*
-            if(hide_mode){
-              exist = find_nodes_south(grid, current_node, left, temp_nodes);
-              exist = Check_Src_Dest(temp_nodes, src_index);
-              }
-      */
       if (!exist) {
         temp_nodes.clear();
         exist = find_nodes_west(grid, current_node, left, temp_nodes);
@@ -756,11 +737,7 @@ bool A_star::find_succsive_parallel_node(Grid &grid, int current_node, int left,
     int exist = 0;
 
     if (mode == 0) {
-      /*      if(hide_mode){
-              exist = find_nodes_east(grid, current_node, right, temp_nodes);
-              exist = Check_Src_Dest(temp_nodes, src_index);
-            }
-      */
+
       if (!exist) {
         temp_nodes.clear();
         exist = find_nodes_north(grid, current_node, right, temp_nodes);