--- a/PlaceRouteHierFlow/router/GcellDetailRouter.cpp
+++ b/PlaceRouteHierFlow/router/GcellDetailRouter.cpp
@@ -1,5 +1,4 @@
 #include "GcellDetailRouter.h"
-
 #include "spdlog/spdlog.h"
 
 GcellDetailRouter::GcellDetailRouter(){