--- a/PlaceRouteHierFlow/router/GcellDetailRouter.h
+++ b/PlaceRouteHierFlow/router/GcellDetailRouter.h
@@ -82,8 +82,8 @@ class GcellDetailRouter : public GcellGlobalRouter {
   // void InsertPathToCand_Start(std::vector<std::vector<RouterDB::Metal> > &physical_path, RouterDB::Segment &temp_seg);
   // void InsertPathToCand_End(std::vector<std::vector<RouterDB::Metal> > &physical_path, RouterDB::Segment &temp_seg);
   void ConvertRect2GridPoints(std::vector<std::vector<RouterDB::point>> &plist, int mIdx, int LLx, int LLy, int URx, int URy);
-  void ConvertRect2GridPoints_set(std::vector<std::set<RouterDB::point, RouterDB::pointXYComp>> &plist, int mIdx, int LLx, int LLy, int URx, int URy);
-  void ConvertRect2GridPoints_Via_set(std::vector<std::set<RouterDB::point, RouterDB::pointXYComp>> &plist, int mIdx, int LLx, int LLy, int URx, int URy);
+  void InactivateRect2GridPoints(int mIdx, int LLx, int LLy, int URx, int URy, Grid& grid);
+  void InactivateRect2GridPoints_Via(int mIdx, int LLx, int LLy, int URx, int URy, bool up, Grid &grid);
   // void CheckOverlapMetals(std::set<RouterDB::SinkData, RouterDB::SinkDataComp> Set_x, std::vector<RouterDB::Net>& nets);
   void Physical_metal_via();
   // void AddMetalToPin();