speedup router using vtune.
Method: change the code implementation, do not change the algorithm and logic

switched_capacotpor_filter
Result: total time 800s -> 216s (3.7x), routing time: 751s -> 169s (4.4x)