@854768750 2022-11-22
@stevenmburns @soneryaldiz I will work on this PR first, before #1199 

The testcase in #1199 switched_capacitor_filter takes 14 minutes each run and is not acceptable for me.
Now I have one code change make it to be 6 minutes. Will have more code change to make it acceptable for me.
Thanks @stevenmburns introducing vtune in #1037