@854768750 2022-11-23
@stevenmburns @soneryaldiz @srini229 This pull request makes some code change to make the router 4.4x faster on switched_capacitor_filter than the master branch. The code does not change the routing algorithm or logic, just change the way the code is implemented.